Assignment 3 - CEE 218X
This assignment uses PUMS and ACS Census data to conduct an analysis of languages spoken and income with the population focus being on Hispanic people in the Bay Area. I do not own any of the course materials for CEE 218X. I do not own any of the data. This assignment is for educational purposes only.