Assignment 6 - CEE 218X
This assignment used PUMS, ACS, and Mapbox, and San Francisco open parcel data to analyze Sonoma County renters' housing costs and parcel availability in San Francisco. I do not own any of the course materials for CEE 218X. I do not own any of this data. This assignment is for educational purposes only.