Assignment 2 - CEE 218X
This assignment involved the use of open source Census ACS and PUMS data to create an equity analysis of Sonoma County educational attainment. Educational mobility for residents of Sonoma County is also analyzed. I do not own the data or course materials for CEE 218X. All of these materials are used for educational purposes only.