Assignment 7 - Part 2
This is the second part of Assignment 7 for CEE 218X that analyzed total emissions from residential and commercial buildings in Sebastopol, CA by using Sonoma County open data and open PG&E data. I do not own any of the course materials for CEE 218X. I do not own any of the data from sources. This page was creating for educational purposes only.