Assignment 7 - Part 1
The is the first part of Assignment 7 for CEE 218X which analyzes the total green house gas emissions (in tons) produced by vehicles commuting to Sonoma County. Mapbox and ACS data were used in this analysis. Assignment 7 is divided into 2 HTML files for rendering purposes. I do not own any of the course materials for CEE 218X. I do not own any of the data. This analysis is for educational purposes only.