Assignment 5 - CEE 218X
This assignment focused on using open source data to analyze disparities in distance to polling places by household in Sonoma County. BayAreaCommunity.org and unboxproject.org provided the data set that was combined with ACS data to prepare this analysis. I do not own any of the course materials for CEE 218X. I do not own any of the data used in the analyses. This page is for educational purposes only.